Change : Toast and Dialog class move to widzets package
1. Moved the toast and progress dialog class to utils -> widzets.
