+++ movielist/app/src/main/java/com/list/movie/hyuck/movielist/widzets/CustomProgressDialog.java
@@ -1,4 +1,4 @@
-package com.list.movie.hyuck.movielist.utils;
+package com.list.movie.hyuck.movielist.widzets;
 
 import android.content.Context;
 import android.graphics.Color;
